backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
groq>=0.9.0
pydantic>=2.7.0
httpx>=0.27.0
python-docx>=1.1.2
PyMuPDF>=1.24.0
python-dotenv>=1.0.1
